backend/services/ai_refinement/requirements.txt
fastapi
openai
uvicorn
pyyaml
pydantic
aiohttp
pytest
python-dotenv

python-jose
passlib[bcrypt]
redis
